chore: Show Suggested Widgets menu in Response tab
